+++ src/renderer/main.tsx
@@ -1,15 +1,15 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import {Store, createStore, Middleware, applyMiddleware} from 'redux';
-import {createLogger} from 'redux-logger';
+import { Store, createStore, Middleware, applyMiddleware, Dispatch } from 'redux';
+import { createLogger } from 'redux-logger';
 import thunkMiddleware from 'redux-thunk';
-import {Provider} from 'react-redux';
-import {ipcRenderer} from 'electron';
+import { Provider } from 'react-redux';
+import { ipcRenderer } from 'electron';
 import ApplicationPage from './containers/ApplicationPage'
-import {newWebAPIClient} from './webapi'
-import {State} from './state';
+import { newWebAPIClient } from './webapi'
+import { State } from './state';
 import * as actions from './actions'
-import {stateReducer} from './reducers';
+import { stateReducer } from './reducers';
 
 
 export function main() {
@@ -65,6 +65,19 @@ export function main() {
     ipcRenderer.on('get-preferences', () => {
         store.dispatch(actions.sendPreferencesToMain());
     });
+
+    document.addEventListener('drop', function (event: any) {
+        event.preventDefault();
+        event.stopPropagation();
+        for (let file of event.dataTransfer.files) {
+            readDroppedFile(file, store.dispatch);
+        }
+    });
+
+    document.addEventListener('dragover', function (event: any) {
+        event.preventDefault();
+        event.stopPropagation();
+    });
 }
 
 function connectWebAPIClient(store: Store<State>) {
@@ -102,5 +115,28 @@ function connectWebAPIClient(store: Store<State>) {
     };
 }
 
+function readDroppedFile(file: File, dispatch: Dispatch<State>) {
+    let opName;
+    if (file.path.endsWith('.nc')) {
+        opName = 'read_netcdf';
+    } else if (file.path.endsWith('.txt')) {
+        opName = 'read_text';
+    } else if (file.path.endsWith('.json')) {
+        opName = 'read_json';
+    } else if (file.path.endsWith('.csv')) {
+        opName = 'read_csv';
+    } else if (file.path.endsWith('.geojson') || file.path.endsWith('.shp') || file.path.endsWith('.gml')) {
+        opName = 'read_geo_data_frame';
+    }
 
+    if (opName) {
+        dispatch(actions.setWorkspaceResource(opName,
+                                              {file: {value: file.path}},
+                                              null,
+                                              false,
+                                              'Reading dropped file'));
+    } else {
+        console.warn('Dropped file of unrecognized type: ', file.path);
+    }
+}